// client/src/enum/ClassStandardEnums.ts
enum ClassStandard {
    FIRST_STANDARD = "1",
    SECOND_STANDARD = "2",
    THIRD_STANDARD = "3",
    FOURTH_STANDARD = "4",
    FIFTH_STANDARD = "5",
    SIXTH_STANDARD = "6",
    SEVENTH_STANDARD = "7",
    EIGHTH_STANDARD = "8",
    NINTH_STANDARD = "9",
    TENTH_STANDARD = "10",
    ELEVENTH_STANDARD = "11",
    TWELFTH_STANDARD = "12",
  }

  export default ClassStandard